make vm a pure interface with docker and inproc impls.
